# Filter out dead threads in runtime metrics